Refactor HotRodStore resource definition and service configurator into same package as all other cache stores.
