+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/XMLAttribute.java
@@ -28,7 +28,6 @@ import java.util.Map;
 
 import org.jboss.as.clustering.controller.Attribute;
 import org.jboss.as.clustering.infinispan.subsystem.remote.ConnectionPoolResourceDefinition;
-import org.jboss.as.clustering.infinispan.subsystem.remote.HotRodStoreResourceDefinition;
 import org.jboss.as.clustering.infinispan.subsystem.remote.RemoteCacheContainerResourceDefinition;
 import org.jboss.as.clustering.infinispan.subsystem.remote.RemoteClusterResourceDefinition;
 import org.jboss.as.clustering.infinispan.subsystem.remote.SecurityResourceDefinition;
